# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST em Go com Clean Architecture para processamento de textos, análise de currículos com Gemini e disparos automáticos.",
        "title": "API Go de Processamento de Texto e Documentos",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/match": {
            "post": {
                "description": "Recebe apenas o currículo do candidato em base64. Gera o embedding vetorial do currículo, executa a busca de similaridade rápida no banco local (chromem-go) e envia as vagas compatíveis para triagem detalhada via LLM (Gemini/DeepSeek), realizando também os disparos automáticos.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Processador"
                ],
                "summary": "Executa match de currículo contra banco de vagas",
                "parameters": [
                    {
                        "description": "Payload contendo apenas o currículo base64",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.matchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultado da busca vetorial e envios automáticos",
                        "schema": {
                            "$ref": "#/definitions/domain.ProcessResult"
                        }
                    },
                    "400": {
                        "description": "Requisição inválida",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro ao extrair, buscar ou enviar candidaturas",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vacancies": {
            "post": {
                "description": "Recebe uma lista de vagas em texto bruto e um delimitador. Divide o texto, gera os embeddings vetoriais via Ollama (em lote) e as salva no banco vetorial local (chromem-go) de forma persistente.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vagas"
                ],
                "summary": "Popula o banco vetorial de vagas",
                "parameters": [
                    {
                        "description": "Payload com as vagas e delimitador",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.populateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vagas salvas e indexadas com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Requisição inválida",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro ao processar e salvar vagas",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vacancies/clear": {
            "post": {
                "description": "Remove todas as vagas e coleções existentes no banco de dados local do chromem-go.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vagas"
                ],
                "summary": "Limpa o banco de dados vetorial de vagas",
                "responses": {
                    "200": {
                        "description": "Banco vetorial de vagas limpo com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno ao limpar banco vetorial",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Detailed error message here"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "domain.FileMetadata": {
            "type": "object",
            "properties": {
                "mime_type": {
                    "type": "string"
                },
                "size_in_bytes": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "domain.Match": {
            "type": "object",
            "properties": {
                "contact_target": {
                    "description": "email address or phone number",
                    "type": "string"
                },
                "contact_type": {
                    "description": "\"email\" or \"whatsapp\"",
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "domain.ProcessResult": {
            "type": "object",
            "properties": {
                "bytes_processed": {
                    "type": "integer"
                },
                "duration_ms": {
                    "type": "integer"
                },
                "file": {
                    "$ref": "#/definitions/domain.FileMetadata"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "items_processed": {
                    "type": "integer"
                },
                "matches": {
                    "description": "matches detected by Gemini",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Match"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "handler.matchRequest": {
            "type": "object",
            "properties": {
                "file_base64": {
                    "type": "string"
                }
            }
        },
        "handler.populateRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "delimiter": {
                    "type": "string"
                }
            }
        }
    }
}